requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
requests
pandas
numpy
scikit-learn
faiss-cpu
jinja2
open-clip-torch